C - encrpt & decrypt using Caesar Cypher Algorithm
Write a C/C++ program to encrypt and decrypt the string using Caesar Cypher Algorithm. While
encrypting the given string, 5 is added to the ASCII value of the characters. Similarly, for decrypting the
string, 5 is subtracted from the ASCII value of the characters to print an original string. Example - Input
String - hello, Encrypted String - mjqqt .